+++ CHANGES.rst
@@ -32,6 +32,7 @@ Changes in 0.18.11 (2021-05-07)
  * MXCallKitAdapter: Update incoming calls if answered from application UI.
  * MXFileStore: Logs all files when a data corruption is detected (to track vector-im/element-ios/issues/4921).
  * MXCallManager: Fix call transfers flow for all types of transfers.
+ * VoIP: Implement asserted identity for calls: MSC3086 (matrix-org/matrix-doc/pull/3086).
 
 🐛 Bugfix
  * MXTools: Fix bad linkification of matrix alias and URL (vector-im/element-ios/issues/4258).